Mexendo na pagina de previsoes, falta estética, mas ja ta funcional
Implements a new page in the dashboard for multi-day price predictions.

This allows users to input a ticker and a number of days to predict future prices, replacing the original single 10-day prediction.

Includes styling updates for input elements.

Refactors price regressor to allow forecasting for an interval of dates, saving results to the database.

Updates the regressor pipeline to predict N days in the future and calculate related metrics. It also introduces menu options for running the regressor and saving the output to the database.